+++ CHANGELOG.md
@@ -5,10 +5,12 @@
     - [Breaking changes](#breaking-changes)
     - [New features](#new-features)
     - [Fixes & maintenance](#fixes--maintenance)
+  - [0.4.1](#041)
+    - [Fixes & maintenance](#fixes--maintenance-1)
   - [0.4.0](#040)
     - [Breaking changes](#breaking-changes-1)
     - [New features](#new-features-1)
-    - [Fixes & maintenance](#fixes--maintenance-1)
+    - [Fixes & maintenance](#fixes--maintenance-2)
   - [0.3.1](#031)
 
 ## Unreleased
@@ -16,6 +18,10 @@
 ### Breaking changes
 
 ### New features
+
+### Fixes & maintenance
+
+## 0.4.1
 
 ### Fixes & maintenance